institutions - autoForm hooks, form style, shared form for insert/update
when form is submitted, the autoForm hook then fires the saving
function, which hides the submit button and shows saving element. on
success, redirected to institution profile page.

admin institution form helper - this is necessary for determining if
update or insert. the session variable use to create this helper is set
in the router