Edit Article - affiliations, data as session variable, route capitalized name
Add/Remove Affiliation updates the data used to generate the article
form template.

Remove: The field ‘affiliations’ and the field ‘authors’ are both
updated

**TODO**: Template variable helper not updating after session variable
update. This is seen on the affiliation number, which uses the index in
the array as the affiliation number.

FUNCTIONS
getAffiliations: used when updating order and when gathering form data
to update doc in db

updateAffiliationsOrder: used in update method of .sortable

TODO: Add affiliations to author, Update article doc