Data Submissions - Article Editing
Adjusted Admin Article Editing so that form template, events, and data
can be used in Data Submissions for article editing

Session variable ’preprocess-article' now determines whether to
preprocess article data before passing to template. Before, we were
just testing for whether session variable ‘article’ existed, but on
Data Submissions this will change.

Data Submission was using a server method to organize the submissions
list, this way we don’t need to client to subscribe to the entire
articles collection. However, this causes problems when changing which
article to edit on Data Submissions. But via the session variable used
to generate the list ‘submissions_list’, we get the article and set the
session variable ‘article’ without ever changing subscription.