XML to JSON - reference PubMed links and different types
PubMed links
———————
If reference in XML has PMID, then link in template.
A template helper, getPmid, is used when any reference does not have a
PMID. This helper will use the method getPubMedId() to query PubMed for
the PMID. If PMID is returned to the helper getPmid, then update the
Session variable.

Different types
——————
Different reference style of journal article or book