XML to JSON - Reference PMID check
For references without PMID in the XML, a template helper is used to
query PubMed and then update the session variable, article-text, to
include PMID. Before including though, verify that the PMID is correct
by searching PubMed using the PMID and seeing if titles match. If
match, then return PMID. Otherwise, return false.