Admin - For Authors. cleanWysiwyg  cleaning global. EdBoard collection publication edits
Same page to add and edit sections. Same form to add and edit sections.
WYSIWYG for section title and content. Two session variables used,
showForm and sectionId. showForm is used for new sections. sectionId is
used to show the form when editing an existing section. Session
variables are set on click event and cleared when template destroyed.

cleanWysiwyg
global search for tags

EdBoard
The role changed in the member doc from a string to an array