Save article & XML parsing error & DOI project
If value empty, it might be that an article was removed from article
type (or whatever is empty). So we want to update the DB to have an
empty value.

XML parsing error
Include XML URL in error message if there was a problem parsing from S3

DOI Project
temporary functions