News - Video & Tags & Interview
Video: Only YouTube right now. Video input is the ID only. If ‘http’ is
included in the ID then the user is alerted. YouTube video verified
real before saving via metho in newsMethods, realYouTubeVideo()

Tags: Simply an array in the news doc. These can have italics,
superscripts and subscripts. Tags can be added or removed but not
rearranged.

Interview: Simple checkbox underneath video to indicate if the video is
an interview.