XML processing before upload
PDF/XML uploader now on separate page as PDF/XML settings page.

If uploading XML, the article form is shown, as well as conflicts with
the DB. There’s a button above the form for the user to verify
uploading the XML. Updating the database is dependent on the users.
After uploading the XML, the user can still save changes to the
database.