Article Form - Using method for validation
Duplicate article check not just for inserting articles anymore. Now
when updating articles, also checking for duplicate (for ex, in case of
title change)

Moved form validation to method, which will also use the duplicate
article check method, articleExistenceCheck(). The result of the
validation method will be and object with either: duplicate, invalid,
or saved set to true. These flags are checked in the event handler,
‘submit form’ within Template.AdminArticleForm.events.

When result.duplicate = true
The duplicate article is returned

When result.invalid = true
Will also contain array of object for inputs found that weren’t valid

When result.saved=true
If article was inserted/update, then _id is returned.