Delete Issue
Deleting an issue will delete all issue information from the database
and remove all issue information for its articles.

User input to confirm deletion is sent to deleteIssue(), which makes
sure that the input === 'DELETE'. If so then, the issue doc is copied
to the issues_deleted collection via deleteIssueDatabase() and the
original doc is removed from the issues collection. Then all articles
in the issue are updated (just issue information removed from article
docs).

TODO: Cover deleting on S3.