Admin XML processing and comparison with the DB
Form inputs now merge data in objects within arrays in article data.
For example, if author missing middle name in XML but in DB. This was
already shown as a conflict, but now the object data is merged and
complete as possible.

XML is always still the default data source, if data available.