Current Issue
Validation to make sure display is also checked. Before updating, if
another issue was set to current then unset. If issue is set to
current, can’t be unchecked. So that there is always a current issue.
The only way to change current issue is by setting another issue to
current.