Altmetric badge embedding
Changed badge from span to div and put into partial template. Badge
partial template has a callback is calling altmetric embed script to
check page again for matches.

Using event listener to set session variable, which is used in template
helper to add class (if badge is visible) so that the title is padded
and not wrapped around the badge